Finished data generation for 1 and 2
Finished the training data. The data is now labeled 0 if i's outside of the triangle and 1 if it's inside of the the triangle. I also already added the noise parameter and let the data depend on that. Only thing still needed for the data generation is the editing of the triangle, but that should not be too difficult and i'll probably do that tomorrow (today? :eyes: )

TODO: make the data generation for the different sized triangles, make the knn model and test, export the amount of misclassifications to excel